requirements/prod.txt
-r base.txt



django-storages



django-redis
channels-redis
python-decouple==3.8

